requirements.txt
pandas
pyarrow
boto3
psycopg2-binary
python-dotenv
SQLAlchemy
